Move stats api to separate file
